requirements.txt
streamlit
google-genai
stanza
pyphen
wordfreq
